Adjust Huffman tree naming
Our Huffman tree implementation turns out to not be very Deflate-specific
anymore, so this general terminology makes more sense. This is a big bikeshed,
but we can probably get away with it this early. Technically it also means that
hot upgrades from 1.0.0 are possible, although this doesn't matter one bit atm.