Add automatic multi-member handling
Forcing the user to handle multi-member files manually is a great way to cause
very rare and subtle truncation errors. Despite changing the recommended API,
this is completely backwards-compatible and doesn't leave deprecated functions
behind. The old allow-overreads-p option is extremely niche now, but it's still
supported for compatibility and because it technically still makes sense.